Add experimental bitmasking code
Using C bitfields seems to generate slower code than manual masking the
bits does.  The code I've added here is not as careful as what the
compiler would have to generate to ensure wrapping is handled correctly.

This patch seems to fix the performance regression seen when benchmarking
palloc/pfree in a tight loop.  Maybe there is some better way we can do
this, but this is still a prototype.