Add central declarations for dlsym()ed symbols
This is in preparation for defaulting to -fvisibility=hidden in extensions,
instead of relying on all symbols in extensions to be exported. For that
exported symbols need to be tagged with PGDLLEXPORT. Most extensions only need
to do so for _PG_init() and functions defined with PG_FUNCTION_INFO_V1. Adding
central declarations with PGDLLEXPORT avoids breaking extensions declaring
_PG_init(), as long as they include fmgr.h before the declaration.

Author: Andres Freund <andres@anarazel.de>
Discussion: https://postgr.es/m/20211101020311.av6hphdl6xbjbuif@alap3.anarazel.de